+++ crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/artcraft/mod.rs
@@ -1,0 +1,6 @@
+pub (super) mod handle_inpaint_artcraft;
+
+mod handle_artcraft_flux_pro_1_inpaint;
+mod handle_artcraft_flux_dev_juggernaut_inpaint;
+mod handle_artcraft_flux_pro_kontext_inpaint;
+mod handle_artcraft_nano_banana_inpaint;
