--- crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/flux_dev_juggernaut_inpaint/mod.rs
@@ -1,2 +1,0 @@
-pub (super) mod handle_flux_dev_juggernaut_inpaint;
-mod handle_flux_dev_juggernaut_inpaint_artcraft;
