--- crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/flux_pro_1_inpaint/mod.rs
@@ -1,2 +1,0 @@
-pub (super) mod handle_flux_pro_1_inpaint;
-mod handle_flux_pro_1_inpaint_artcraft;
