--- crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/flux_pro_kontext_inpaint/mod.rs
@@ -1,2 +1,0 @@
-pub (super) mod handle_flux_pro_kontext_inpaint;
-pub (super) mod handle_flux_pro_kontext_inpaint_artcraft;
