--- crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/gemini_25_flash_inpaint/mod.rs
@@ -1,2 +1,0 @@
-pub (super) mod handle_gemini_25_flash_inpaint;
-pub (super) mod handle_gemini_25_flash_inpaint_artcraft;
