+++ crates/desktop/artcraft/src/core/commands/enqueue/image_inpaint/mod.rs
@@ -1,5 +1,4 @@
-mod flux_dev_juggernaut_inpaint;
-mod flux_pro_1_inpaint;
-mod flux_pro_kontext_inpaint;
-mod gemini_25_flash_inpaint;
+pub (super) mod inpaint_models;
 pub mod enqueue_image_inpaint_command;
+
+mod artcraft;
